requirements.txt
fastapi
uvicorn
pillow
pytesseract
motor
pymongo
google-generativeai
python-dotenv
